refactor: fixed conflicts and separate by entities resolvers and add foundational testing resolvers
