+++ src/gql/resolvers/ads/ad_resolver.test.ts
