+++ src/gql/resolvers/index.ts
@@ -1,7 +1,7 @@
 import { Resolvers } from '../types/typedSchema'
-import authenticationResolvers from './authenticationResolver';
-import partnerResolvers from './partnerResolver';
-import advertisingResolvers from './advertisingResolver';
+import authenticationResolvers from './auth/authenticationResolver'
+import partnerResolvers from './partners/partnerResolver'
+import advertisingResolvers from './ads/advertisingResolver'
 
 export const resolvers: Resolvers = {
   Query: {
