+++ src/gql/resolvers/partners/partnerResolver.ts
@@ -1,0 +1,26 @@
+import { Container } from 'typedi'
+import { Resolvers } from '../../types/typedSchema'
+import PartnerModel from '../../../db/models/partner'
+import { AuthAction } from '../../../actions/authAction'
+
+Container.set('PartnerModel', PartnerModel)
+const authAction = Container.get<AuthAction>(AuthAction)
+
+
+const partnerResolvers: Resolvers = {
+  Query: {
+    // Add queries for partners
+  },
+  Mutation: {
+    registerPartner: async (_: any, { input }) => {
+      try {
+        return await authAction.registerPartner(input)
+      } catch (e) {
+        console.error('Registration error:', e)
+        throw e
+      }
+    }
+  }
+}
+
+export default partnerResolvers
